+++ CHANGELOG.md
@@ -4,4 +4,4 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 - Start of new changelog.
-- Fixed GM SB XRAY button using stored order ID when none found in email.
+- Added Reset Extension button in the popup to clear storage and reload.